refactor: move code from main to lib
